Remove unnecessary parameters from the readme.txt
"Requires at least: 6.0" is not needed in your "readme.txt" file - Redundant as of WordPress 5.8. Use only the plugin's main PHP file "admin-debug-tools.php" for this declaration.
"Requires PHP: 8.0" is not needed in your "readme.txt" file - Redundant as of WordPress 5.8. Use only the plugin's main PHP file "admin-debug-tools.php" for this declaration.

Please check the documentation for the correct parameters for each file:
- For your readme file: https://developer.wordpress.org/plugins/wordpress-org/how-your-readme-txt-works/
- For your main PHP file: https://developer.wordpress.org/plugins/plugin-basics/header-requirements/